milky_way/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.1
django-phonenumber-field==7.2.0
idna==3.4
loguru==0.7.2
phonenumbers==8.13.22
Pillow==10.1.0
psycopg2==2.9.9
python-barcode==0.15.1
reportlab==4.0.6
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
